aivoice/aivoice/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
av==14.1.0
base==0.0.0
basicsr==1.4.2
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.2.0
bottle==0.12.25
bottle-websocket==0.2.9
breadability==0.1.20
Brotli==1.1.0
bson==0.5.10
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
channels==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
clipman==3.3.1
cloudpathlib==0.20.0
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.7
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
ctranslate2==4.5.0
curl_cffi==0.7.4
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.18.2
cycler==0.12.1
cymem==2.0.11
daphne==4.1.2
darkdetect==0.8.0
dbus-next==0.2.3
decorator==4.4.2
deep-translator==1.11.4
deepface==0.0.93
diffusers==0.32.2
distlib==0.3.8
distro==1.9.0
Django==5.1.4
docopt==0.6.2
easydict==1.13
edge-tts==7.0.0
Eel==0.16.0
emoji==2.14.1
facenet-pytorch==2.6.0
facexlib==0.3.0
fake-useragent==2.0.3
fastapi==0.115.11
faster-whisper==1.1.1
fer==22.5.1
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.15.4
filterpy==1.4.5
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
g4f==0.4.4.4
gast==0.6.0
gdown==5.2.0
gevent==24.2.1
gevent-websocket==0.10.1
gfpgan==1.3.8
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.30.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
gradio==5.20.0
gradio_client==1.7.2
greenlet==3.1.1
groovy==0.1.2
groq==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hyperframe==5.2.0
hyperlink==21.0.0
HyperPyYAML==1.2.2
idna==2.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.6.1
incremental==24.7.2
insightface==0.2.1
instaloader==4.14
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
keyboard==0.13.5
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.6.2
loguru==0.7.2
lxml==5.3.0
lz4==4.4.3
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
mistune==3.1.1
ml-dtypes==0.4.1
more-itertools==10.6.0
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
mss==9.0.2
mtcnn==1.0.0
multidict==6.1.0
murmurhash==1.0.12
namex==0.0.8
narwhals==1.29.0
nest-asyncio==1.6.0
NetHyTech-MCV==1.1
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.60.0
numpy==1.26.4
ollama==0.4.7
onnx==1.17.0
onnxruntime==1.20.1
openai==1.61.0
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.2
playsound==1.3.0
playwright==1.49.1
pooch==1.8.2
preshed==3.0.9
primp==0.11.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pygrabber==0.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMatting==1.1.13
PyMsgBox==1.0.9
pynput==1.7.7
pyOpenSSL==25.0.0
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.5.4
PyRect==0.2.0
pyreqwest_impersonate==0.5.3
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pyttsx3==2.98
pytube==15.0.0
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.3
pywinauto==0.6.8
PyYAML==6.0.2
realesrgan==0.3.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
rembg==2.0.61
requests==2.32.3
requests-toolbelt==1.0.0
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.9.9
safehttpx==0.1.6
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.28.1
semantic-version==2.10.0
sentencepiece==0.2.0
service-identity==24.2.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.2
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
spotipy==2.25.0
sqlparse==0.5.3
srsly==2.5.1
srt==3.5.3
starlette==0.46.0
streamlit==1.42.2
sumy==0.11.0
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.20.0a20250302
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.3.0
textblob==0.19.0
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.8.0
tls-client==1.0.1
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0+cpu
torchaudio==2.6.0
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.2
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.0
virtualenv==20.26.3
vosk==0.3.45
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
webrtcvad==2.0.10
webscout==6.9
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
whichcraft==0.6.1
whisper==1.1.10
wikipedia==1.4.0
win32-setctime==1.1.0
wrapt==1.17.0
wsproto==1.2.0
xformers==0.0.29.post3
yapf==0.43.0
yarl==1.18.3
yaspin==3.1.0
zipp==3.21.0
zope.event==5.0
zope.interface==6.3
